+++ NEWS.md
@@ -1,6 +1,9 @@
 # devtools 1.7.0
 
 ## Improve reverse dependency checking
+* `use_coveralls()` allows you to easily add coveralls support to a package. (#680, @jimhester)
+
+* Fixed scoping issues with `unzip()`.
 
 Devtools now supports a new and improved style of revdep checking with `use_revdep()`. This creates a new directory called `revdep` which contains a `check.R` template. Run this template to check all reverse dependencies, and save summarised results to `check/summary.md`. You can then check this file into git, making it much easier to track how reverse dependency results change between versions. The documentation for `revdep_check()` is much improved, and should be more useful (#635)